Refactor codes by EXCHANGE logic
