+++ static/js/messageTypes.js
@@ -1,0 +1,9 @@
+export const MESSAGE_TYPES = {
+    ACTIVATE: 'ACTIVATE',
+    PULL: 'PULL',
+    PUSH: 'PUSH',
+    FORWARD: 'FORWARD',
+    ARRANGE: 'ARRANGE',
+    FETCH: 'FETCH',
+    EXCHANGE: 'EXCHANGE',
+};
